+++ sdk/Request/CalibrationStartRequest.cs
@@ -25,4 +25,19 @@ namespace EyeTribe.ClientSdk.Request
         public CalibrationStartRequestValues Values { set; get; }
     }
 
+    internal class CalibrationStartRequestValues
+    {
+        public CalibrationStartRequestValues()
+        {
+            PointCount = 9; // default number of calibration points
+        }
+
+        public CalibrationStartRequestValues(int numPoints)
+        {
+            PointCount = numPoints;
+        }
+
+        [JsonProperty(PropertyName = Protocol.CALIBRATION_POINT_COUNT)]
+        public int PointCount { set; get; }
+    }
 }